requirements.txt
livekit-agents
livekit-agents[openai]~=1.0
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-silero
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn